zimmer-backend/requirements.txt
# Zimmer AI Platform Backend Requirements
# Python 3.11+ required

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib==1.3.1

# HTTP & API
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Environment & Configuration
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Caching
redis==5.0.1

# Payment Integration
zarinpal==1.0.0

# OpenAI Integration
openai==1.3.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production
gunicorn==21.2.0

# System monitoring
psutil

# Email validation
email-validator

# JWT
PyJWT